refactor: delegate cruiser state input actions to service
